Continue to refactor to MVVM
