+++ app/build.gradle
@@ -30,8 +30,9 @@ dependencies {
     compile 'com.android.support:design:23.4.0'
     compile 'com.android.support:recyclerview-v7:23.4.0'
     compile 'com.android.support:cardview-v7:23.4.0'
-    // jsoup HTML parser library @ http://jsoup.org/
-    compile 'org.jsoup:jsoup:1.9.2'
     // GSON for TinyDB
     compile 'com.google.code.gson:gson:2.8.0'
+    // Butterknife
+    compile 'com.jakewharton:butterknife:8.5.1'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:8.5.1'
 }
