+++ app/src/main/java/io/ericlee/illinilaundry/View/Adapters/ViewPagerAdapter.java
@@ -1,12 +1,11 @@
-package io.ericlee.illinilaundry.Adapters;
+package io.ericlee.illinilaundry.View.Adapters;
 
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentStatePagerAdapter;
 
-import io.ericlee.illinilaundry.Tabs.FragmentAlarms;
-import io.ericlee.illinilaundry.Tabs.FragmentAll;
-import io.ericlee.illinilaundry.Tabs.FragmentBookmarks;
+import io.ericlee.illinilaundry.View.Fragments.FragmentAll;
+import io.ericlee.illinilaundry.View.Fragments.FragmentBookmarks;
 
 public class ViewPagerAdapter extends FragmentStatePagerAdapter {
     int mNumOfTabs;
@@ -21,12 +20,9 @@ public class ViewPagerAdapter extends FragmentStatePagerAdapter {
 
         switch (position) {
             case 0:
-                FragmentAlarms tab0 = new FragmentAlarms();
-                return tab0;
-            case 1:
                 FragmentAll tab1 = new FragmentAll();
                 return tab1;
-            case 2:
+            case 1:
                 FragmentBookmarks tab2 = new FragmentBookmarks();
                 return tab2;
             default:
