+++ app/src/main/res/layout/card_dorms.xml
@@ -4,10 +4,10 @@
     <data>
         <variable
             name="viewModel"
-            type="io.ericlee.illinilaundry.Adapters.DormCardAdapter" />
+            type="io.ericlee.illinilaundry.ViewModel.DormViewModel" />
     </data>
 
-    <android.support.v7.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
+    <android.support.v7.widget.CardView
         xmlns:card_view="http://schemas.android.com/apk/res-auto"
         android:id="@+id/card_view"
         android:layout_width="match_parent"
@@ -19,8 +19,7 @@
         card_view:cardCornerRadius="2dp"
         card_view:cardElevation="2dp">
 
-        <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-            xmlns:card_view="http://schemas.android.com/apk/res-auto"
+        <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:layout_margin="0dp"
@@ -45,58 +44,11 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:gravity="center_horizontal"
-                android:singleLine="true"
-                android:text="placeholder"
+                android:maxLines="1"
+                android:text="@{viewModel.name}"
                 android:textColor="@color/white"
                 android:textSize="12sp" />
 
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="30dp"
-                android:baselineAligned="false"
-                android:orientation="horizontal"
-                android:weightSum="4">
-
-                <ImageView
-                    android:id="@+id/imageWasher"
-                    android:layout_width="0dp"
-                    android:layout_height="match_parent"
-                    android:layout_weight="1"
-                    android:padding="4dp"
-                    android:scaleType="fitCenter"
-                    android:src="@drawable/washer" />
-
-                <TextView
-                    android:id="@+id/txtWasher"
-                    android:layout_width="0dp"
-                    android:layout_height="match_parent"
-                    android:layout_marginTop="1dp"
-                    android:layout_weight="1"
-                    android:gravity="start|center"
-                    android:text="0"
-                    android:textColor="@color/white"
-                    android:textSize="12sp" />
-
-                <TextView
-                    android:id="@+id/txtDrier"
-                    android:layout_width="0dp"
-                    android:layout_height="fill_parent"
-                    android:layout_marginTop="1dp"
-                    android:layout_weight="1"
-                    android:gravity="center|end"
-                    android:text="0"
-                    android:textColor="@color/white"
-                    android:textSize="12sp" />
-
-                <ImageView
-                    android:id="@+id/imageDrier"
-                    android:layout_width="0dp"
-                    android:layout_height="match_parent"
-                    android:layout_weight="1"
-                    android:padding="4dp"
-                    android:scaleType="fitCenter"
-                    android:src="@drawable/dryer" />
-            </LinearLayout>
         </LinearLayout>
     </android.support.v7.widget.CardView>
 </layout>
