+++ app/src/main/res/values/dimens.xml
@@ -10,5 +10,5 @@
     <dimen name="toolbar_height">56dp</dimen>
     <dimen name="toolbar_tab_height">104dp</dimen>
 
-    <dimen name="item_offset">8dp</dimen>
+    <dimen name="item_offset">6dp</dimen>
 </resources>
